Add a new module: 'welfare'
This is part of a larger refactoring effort of the remind core.
Here, code related to the social welfare functions has been moved
from the core to the new module. Some renaming was required. For
now, only one module realization exists, 'utilitarian', which
corresponds one to one to what was in the core before. This
refactoring exercises has no effect on the results!