Add new module: 'trade'
This is part of a larger refactoring effort of the remind core. Here, code related to the trade of permits and primary energy carriers is moved from the core to the new module. Some renaming is required. For now only one module realization exists, 'standard', which corresponds exactly to what was in the core before. The results are not effected!