Add check of R library to guarantee package loads
At the start of R session, a check of the R library is performed to make sure that no packages are being updated. If so, i.e. if a folder beginning with "00LOCK" is found, the library path is switched to the latest snapshot. This avoids the failure of on-cluster runs caused by package updates.