New trade model in network_trade realization.
The network_trade realization, which was previously just a duplicate of the se_trade
model, now has code belonging to the new trade model in its sets, delcarations,
datainput, and postsolve GAMS files. This code is also loaded by an external toy model,
which can be used for testing purposes of the trade model.