make the q32_flexAdj equation more error-proof
pm_SEPrice is only used for the calculation of vm_flexAdj as long as the pm_SEPrice is below 230$/MWh (equal to 2T$/TWa), otherwise 230$/MWh is used as maximum value. 
In the real world, having electricity prices > such a value for a 5 year period seems hardly imaginable. 
(the energy crisis in the EU forced power prices up to 200-400 €/MWh for a year, but then they dropped down to below 200€/MWh)